U1_Project1_Arjun Version 2.0
Got the ball to randomize its movements a little bit more now, by having to increase the speed of the ball. If any problems, there are some times that the ball would kind of stick to the racket for a second. If that happens, move the racket just a little bit to continue playing.